completed testing for all endpoints

+++ routes/tags.js
@@ -29,7 +29,7 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
-  const { userId } = req.user.id;
+  const userId  = req.user.id;
 
   console.log(userId, id);
 
@@ -39,8 +39,8 @@ router.get('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Tag.find({ _id: id, userId })
-    .then(result => {
+  Tag.findOne({ _id: id, userId })
+    .then((result) => {
       if (result) {
         res.json(result);
       } else {
@@ -83,6 +83,8 @@ router.put('/:id', (req, res, next) => {
   const { name } = req.body;
   const  userId  = req.user.id;
 
+  console.log('i am the put request', userId);
+
   /***** Never trust users - validate input *****/
   if (!name) {
     const err = new Error('Missing `name` in request body');
@@ -96,9 +98,7 @@ router.put('/:id', (req, res, next) => {
     return next(err);
   }
 
-  console.log({ _id: id, userId });
-
-  Tag.findOneAndUpdate({ _id: id, userId:userId }, { name }, { new: true })
+  Tag.findOneAndUpdate({ _id: id, userId }, { name }, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -119,12 +119,19 @@ router.put('/:id', (req, res, next) => {
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
   const userId = req.user.id;
-  const tagRemovePromise = Tag.findOneAndRemove({_id: id, userId});
-  // const tagRemovePromise = Tag.remove({ _id: id }); // NOTE **underscore** _id
+
+  /***** Never trust users - validate input *****/
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
+
+  const tagRemovePromise = Tag.findOneAndRemove({ _id: id, userId });
 
   const noteUpdatePromise = Note.updateMany(
-    { 'tags': id, },
-    { '$pull': { 'tags': id } }
+    { tags: id, userId },
+    { $pull: { tags: id } }
   );
 
   Promise.all([tagRemovePromise, noteUpdatePromise])
@@ -135,6 +142,5 @@ router.delete('/:id', (req, res, next) => {
       next(err);
     });
 
-});
-
+  });
 module.exports = router;
